reinit
