+++ languages/java/oso/src/main/java/com/osohq/oso/Oso.java
@@ -5,16 +5,12 @@ import java.util.HashSet;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.List;
-import java.util.ArrayList;
 import java.util.stream.Collectors;
 import org.json.JSONObject;
-import org.json.JSONArray;
 
 public class Oso extends Polar {
-  private boolean polarRolesEnabled;
   public Oso() {
     super();
-    this.polarRolesEnabled = false;
   }
 
   /**
@@ -97,38 +93,6 @@ public class Oso extends Polar {
               }
             })
         .collect(Collectors.toCollection(HashSet::new));
-  }
-
-  public void enableRoles() throws Exceptions.OsoException {
-    if (polarRolesEnabled) return;
-    ffiPolar.enableRoles();
-
-    List<List<HashMap<String, Object>>> allResults = new ArrayList<List<HashMap<String, Object>>>();
-
-    Ffi.Query nextQuery = ffiPolar.nextInlineQuery();
-    for (; nextQuery != null; nextQuery = ffiPolar.nextInlineQuery()) {
-      Host dupHost = host.clone();
-      dupHost.acceptExpression = true;
-      Query query = new Query(nextQuery, dupHost, Map.of());
-      if (!query.hasMoreElements()) {
-        throw new Exceptions.InlineQueryFailedError(nextQuery.source());
-      } else {
-        allResults.add(query.results());
-      }
-    }
-
-    for (List<HashMap<String, Object>> results : allResults) {
-      for (HashMap<String, Object> result : results) {
-        HashMap<String, Object> bs = new HashMap<String, Object>();
-        result.forEach((k, v) -> {
-          bs.put(k, host.toPolarTerm(v));
-        });
-        result.put("bindings", bs);
-      }
-    }
-
-    ffiPolar.validateRolesConfig(new JSONArray(allResults).toString());
-    this.polarRolesEnabled = true;
   }
 
   public static void main(String[] args) throws Exceptions.OsoException, IOException {
